Move the thunks to the correct location
