+++ web/src/core/usecases/s3ProfilesManagement/decoupledLogic/resolveTemplatedStsRole.ts
@@ -1,40 +1,31 @@
 import type { DeploymentRegion } from "core/ports/OnyxiaApi";
 import { id } from "tsafe/id";
-import type { LocalizedString } from "ui/i18n";
 import { z } from "zod";
 import { getValueAtPath } from "core/tools/Stringifyable";
-import { type S3UriPrefixObj, parseS3UriPrefix } from "core/tools/S3Uri";
 
-export type ResolvedTemplateBookmark = {
-    title: LocalizedString;
-    description: LocalizedString | undefined;
-    tags: LocalizedString[];
-    s3UriPrefixObj: S3UriPrefixObj;
-    forProfileNames: string[];
+export type ResolvedTemplateStsRole = {
+    roleARN: string;
+    roleSessionName: string;
+    profileName: string;
 };
 
-export async function resolveTemplatedBookmark(params: {
-    bookmark_region: DeploymentRegion.S3Next.S3Profile.Bookmark;
+export async function resolveTemplatedStsRole(params: {
+    stsRole_region: DeploymentRegion.S3Profile.StsRole;
     getDecodedIdToken: () => Promise<Record<string, unknown>>;
-}): Promise<ResolvedTemplateBookmark[]> {
-    const { bookmark_region, getDecodedIdToken } = params;
+}): Promise<ResolvedTemplateStsRole[]> {
+    const { stsRole_region, getDecodedIdToken } = params;
 
-    if (bookmark_region.claimName === undefined) {
+    if (stsRole_region.claimName === undefined) {
         return [
-            id<ResolvedTemplateBookmark>({
-                s3UriPrefixObj: parseS3UriPrefix({
-                    s3UriPrefix: bookmark_region.s3UriPrefix,
-                    strict: true
-                }),
-                title: bookmark_region.title,
-                description: bookmark_region.description,
-                tags: bookmark_region.tags,
-                forProfileNames: bookmark_region.forProfileNames
+            id<ResolvedTemplateStsRole>({
+                roleARN: stsRole_region.roleARN,
+                roleSessionName: stsRole_region.roleSessionName,
+                profileName: stsRole_region.profileName
             })
         ];
     }
 
-    const { claimName, excludedClaimPattern, includedClaimPattern } = bookmark_region;
+    const { claimName, excludedClaimPattern, includedClaimPattern } = stsRole_region;
 
     const decodedIdToken = await getDecodedIdToken();
 
@@ -106,33 +97,10 @@ export async function resolveTemplatedBookmark(params: {
             const substituteTemplateString = (str: string) =>
                 str.replace(/\$(\d+)/g, (_, i) => match[parseInt(i)] ?? "");
 
-            const substituteLocalizedString = (
-                locStr: LocalizedString
-            ): LocalizedString => {
-                if (typeof locStr === "string") {
-                    return substituteTemplateString(locStr);
-                }
-                return Object.fromEntries(
-                    Object.entries(locStr)
-                        .filter(([, value]) => value !== undefined)
-                        .map(([lang, value]) => [lang, substituteTemplateString(value)])
-                );
-            };
-
-            return id<ResolvedTemplateBookmark>({
-                s3UriPrefixObj: parseS3UriPrefix({
-                    s3UriPrefix: substituteTemplateString(bookmark_region.s3UriPrefix),
-                    strict: true
-                }),
-                title: substituteLocalizedString(bookmark_region.title),
-                description:
-                    bookmark_region.description === undefined
-                        ? undefined
-                        : substituteLocalizedString(bookmark_region.description),
-                tags: bookmark_region.tags.map(tag => substituteLocalizedString(tag)),
-                forProfileNames: bookmark_region.forProfileNames.map(profileName =>
-                    substituteTemplateString(profileName)
-                )
+            return id<ResolvedTemplateStsRole>({
+                roleARN: substituteTemplateString(stsRole_region.roleARN),
+                roleSessionName: substituteTemplateString(stsRole_region.roleSessionName),
+                profileName: substituteTemplateString(stsRole_region.profileName)
             });
         })
         .filter(x => x !== undefined);
