+++ web/src/core/usecases/s3ProfilesManagement/selectors.ts
@@ -90,6 +90,10 @@ const ambientS3Profile = createSelector(
     }
 );
 
+export const protectedSelectors = {
+    resolvedTemplatedBookmarks
+};
+
 export const selectors = {
     isS3ExplorerEnabled,
     s3Profiles,
